soc/qualcomm/common: Include platform subtype in platform ID
Include the platform subtype read from the CDT in the platform ID to
differentiate variants sharing the same platform type

Note: This change is internal to qualcomm specific platforms (CRD/QCB)

Test=Build and boot google/bluey

Reviewed-on: https://review.coreboot.org/c/coreboot/+/94267
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>

+++ src/soc/qualcomm/common/include/soc/cdt.h
@@ -35,7 +35,7 @@ struct platform_id_cdt {
 };
 
 ssize_t cdt_read(void *buffer, size_t buffer_size);
-uint8_t cdt_get_platform_id(void);
+uint16_t cdt_get_platform_id(void);
 uint32_t cdt_get_hw_version(void);
 
 #endif /* _SOC_QUALCOMM_CDT_H_ */
